+++ examples/workflows/low_code_rag/main.tf
@@ -10,7 +10,7 @@ terraform {
   required_providers {
     datarobot = {
       source  = "datarobot-community/datarobot"
-      version = "0.0.20"
+      version = "0.0.22"
     }
   }
 }