+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.22
+
+FEATURES:
+
+- Fix Custom Model runtime parameter ordering
+
+## 0.0.21
+
+FEATURES:
+
+- Add Custom Model training data and class labels
+
 ## 0.0.20
 
 FEATURES: